requirements.txt
torch
torchvision
torchaudio
torchstain
albumentations
scikit-learn
pandas
numpy
matplotlib
seaborn
pillow
tqdm
pytest
